fix: use events to update price in subgraph
Previously, price was checked via `eth_call` upon creating a new Transaction entity.

The problem is that calls see the blockchain state after all the transactions in the
current block have been executed. This means if there was a Liquity TX mined in the
same block as a (Chainlink) price update, but ordered _before_ the price update, the
subgraph would prematurely use the updated price to calculate ICRs while processing
the Liquity TX.

Instead of calling the PriceFeed, use its event `LastGoodPriceUpdated`, which is
correctly ordered with regard to other Liquity events.

Fixes #581.